Moved validate questionnaire to QuestionnaireResponseValidator.kt and removed the need to use questionnaire Response for validation.
