+++ datacapture/src/main/java/com/google/android/fhir/datacapture/validation/QuestionnaireResponseValidator.kt
@@ -59,4 +59,84 @@ object QuestionnaireResponseValidator {
     }
     return linkIdToValidationResultMap
   }
+
+  /**
+   * Traverse (DFS) through the list of questionnaire items and the list of questionnaire response
+   * items and check if the linkId of the matching pairs of questionnaire item and questionnaire
+   * response item are equal.
+   */
+  fun validateQuestionnaireResponseItems(
+    questionnaireResponseInputItemList:
+      List<QuestionnaireResponse.QuestionnaireResponseItemComponent>,
+    questionnaireItemList: List<Questionnaire.QuestionnaireItemComponent>
+  ) {
+    val questionnaireResponseInputItemListIterator = questionnaireResponseInputItemList.iterator()
+    val questionnaireItemListIterator = questionnaireItemList.iterator()
+
+    while (questionnaireResponseInputItemListIterator.hasNext()) {
+      // TODO: Validate type and item nesting within answers for repeated answers
+      // https://github.com/google/android-fhir/issues/286
+      val questionnaireResponseInputItem = questionnaireResponseInputItemListIterator.next()
+      if (questionnaireItemListIterator.hasNext()) {
+        val questionnaireItem = questionnaireItemListIterator.next()
+        if (!questionnaireItem.linkId.equals(questionnaireResponseInputItem.linkId)) {
+          throw IllegalArgumentException(
+            "Mismatching linkIds for questionnaire item ${questionnaireItem.linkId} and " +
+              "questionnaire response item ${questionnaireResponseInputItem.linkId}"
+          )
+        }
+        if (questionnaireResponseInputItem.hasAnswer() &&
+            questionnaireItem.type != Questionnaire.QuestionnaireItemType.GROUP
+        ) {
+          if (!questionnaireItem.repeats && questionnaireResponseInputItem.answer.size > 1) {
+            throw IllegalArgumentException(
+              "Multiple answers in ${questionnaireResponseInputItem.linkId} and repeats false in " +
+                "questionnaire item ${questionnaireItem.linkId}"
+            )
+          }
+          questionnaireResponseInputItem.answer.forEachIndexed {
+            index,
+            questionnaireResponseItemAnswerComponent ->
+            if (questionnaireResponseItemAnswerComponent.hasValue()) {
+              when (questionnaireItem.type) {
+                Questionnaire.QuestionnaireItemType.BOOLEAN,
+                Questionnaire.QuestionnaireItemType.DECIMAL,
+                Questionnaire.QuestionnaireItemType.INTEGER,
+                Questionnaire.QuestionnaireItemType.DATE,
+                Questionnaire.QuestionnaireItemType.DATETIME,
+                Questionnaire.QuestionnaireItemType.TIME,
+                Questionnaire.QuestionnaireItemType.STRING,
+                Questionnaire.QuestionnaireItemType.URL ->
+                  if (!questionnaireResponseItemAnswerComponent
+                      .value
+                      .fhirType()
+                      .equals(questionnaireItem.type.toCode())
+                  ) {
+                    throw IllegalArgumentException(
+                      "Type mismatch for linkIds for questionnaire item ${questionnaireItem.linkId} and " +
+                        "questionnaire response item ${questionnaireResponseInputItem.linkId}"
+                    )
+                  }
+                else -> Unit // Check type for primitives only
+              }
+            }
+            validateQuestionnaireResponseItems(
+              questionnaireResponseItemAnswerComponent.item,
+              questionnaireItem.item
+            )
+          }
+        } else if (questionnaireResponseInputItem.hasItem()) {
+          validateQuestionnaireResponseItems(
+            questionnaireResponseInputItem.item,
+            questionnaireItem.item
+          )
+        }
+      } else {
+        // Input response has more items
+        throw IllegalArgumentException(
+          "No matching questionnaire item for questionnaire response item ${questionnaireResponseInputItem.linkId}"
+        )
+      }
+    }
+  }
 }
